Change bank count data size for mapper 4

+++ src/mapper_004.h
@@ -20,8 +20,8 @@ private:
     uint8_t bank_select_ = 0;
     uint8_t prg_bank_mode_ = 0;
     uint8_t chr_bank_mode_ = 0;
-    uint8_t prg_bank_count_ = 0;
-    uint8_t chr_bank_count_ = 0;
+    uint16_t prg_bank_count_ = 0;
+    uint16_t chr_bank_count_ = 0;
 
     uint8_t mirroring_ = 0;
     uint8_t irq_counter_ = 0;
